Refactored common reducers into its own
