+++ assets/agenda/reducers.js
@@ -35,12 +35,12 @@ import {
 } from './actions';
 
 import { RENDER_MODAL, CLOSE_MODAL } from 'actions';
-import { modalReducer } from 'reducers';
 
 import { get, isEmpty } from 'lodash';
 import { toggleValue } from 'utils';
 import { EXTENDED_VIEW } from 'wire/defaults';
-import { getMaxVersion } from 'wire/utils';
+
+import { defaultReducer } from '../reducers';
 
 const initialState = {
     items: [],
@@ -63,6 +63,7 @@ const initialState = {
     newItemsData: null,
     newItemsByTopic: {},
     readItems: {},
+    resultsFiltered: false,
     agenda: {
         navigations: [],
         activeNavigation: null,
@@ -70,10 +71,9 @@ const initialState = {
         createdFilter: {},
         activeView: EXTENDED_VIEW,
         activeTopic: null,
+        activeDate: Date.now(),
+        activeGrouping: 'day',
     },
-    resultsFiltered: false,
-    activeDate: Date.now(),
-    activeGrouping: 'day',
 };
 
 function recieveItems(state, data) {
@@ -96,26 +96,6 @@ function recieveItems(state, data) {
     };
 }
 
-function getReadItems(state, item) {
-    const readItems = Object.assign({}, state.readItems);
-
-    if (item) {
-        readItems[item._id] = getMaxVersion(readItems[item._id], item.version);
-    }
-
-    return readItems;
-}
-
-function updateItemActions(state, items, action) {
-    const itemsById = Object.assign({}, state.itemsById);
-
-    items.map((item) => {
-        itemsById[item] = Object.assign({}, itemsById[item]);
-        itemsById[item][action] = (itemsById[item][action] || []).concat([state.user]);
-    });
-
-    return itemsById;
-}
 
 function _agendaReducer(state, action) {
     switch (action.type) {
@@ -178,6 +158,14 @@ function _agendaReducer(state, action) {
             activeView: action.view,
         };
 
+    case SELECT_DATE:
+        return {
+            ...state,
+            selectedItems: [],
+            activeDate: action.dateString,
+            activeGrouping: action.grouping || 'day',
+        };
+
     default:
         return state;
     }
@@ -186,94 +174,41 @@ function _agendaReducer(state, action) {
 export default function agendaReducer(state = initialState, action) {
     switch (action.type) {
 
-    case SET_ITEMS: {
-        const itemsById = {};
-        const items = [];
-
-        action.items.forEach((item) => {
-            if (!itemsById[item._id]) {
-                itemsById[item._id] = item;
-                items.push(item._id);
-            }
-        });
-
-        return {
-            ...state,
-            itemsById,
-            items,
-        };
-    }
-
+    case SET_ITEMS:
     case SET_ACTIVE:
-        return {
-            ...state,
-            activeItem: action.item || null,
-        };
-
-    case PREVIEW_ITEM: {
-        const readItems = getReadItems(state, action.item);
-
-        return {
-            ...state,
-            readItems,
-            previewItem: action.item ? action.item._id : null,
-        };
-    }
-
-    case OPEN_ITEM:{
-        const readItems = getReadItems(state, action.item);
-
-        const itemsById = Object.assign({}, state.itemsById);
-        if (action.item) {
-            itemsById[action.item._id] = action.item;
-        }
-
-        return {
-            ...state,
-            readItems,
-            itemsById,
-            openItem: action.item || null,
-        };
-    }
-
-
+    case PREVIEW_ITEM:
+    case OPEN_ITEM:
     case SET_QUERY:
-        return {...state, query: action.query, activeItem: null};
-
-    case SET_EVENT_QUERY:
-        return {...state, queryId: action.query, activeItem: null};
-
     case QUERY_ITEMS:
-        return {...state, isLoading: true, totalItems: null, activeQuery: state.query};
+    case RECIEVE_ITEM:
+    case RECIEVE_NEXT_ITEMS:
+    case RENDER_MODAL:
+    case CLOSE_MODAL:
+    case ADD_TOPIC:
+    case TOGGLE_SELECTED:
+    case SELECT_ALL:
+    case SELECT_NONE:
+    case SHARE_ITEMS:
+    case DOWNLOAD_ITEMS:
+    case COPY_ITEMS:
+    case PRINT_ITEMS:
+    case BOOKMARK_ITEMS:
+    case REMOVE_BOOKMARK:
+    case SET_NEW_ITEMS_BY_TOPIC:
+    case REMOVE_NEW_ITEMS:
+    case SET_NEW_ITEMS:
+    case START_LOADING:
+    case SET_TOPICS:
+    case SET_STATE:
+        return defaultReducer(state || initialState, action);
+
 
     case RECIEVE_ITEMS:
         return recieveItems(state, action.data);
 
-    case RECIEVE_ITEM: {
-        const itemsById = Object.assign({}, state.itemsById);
-        itemsById[action.data._id] = action.data;
-        return  {...state, itemsById};
-    }
+    case SET_EVENT_QUERY:
+        return {...state, queryId: action.query, activeItem: null};
 
-    case RECIEVE_NEXT_ITEMS: {
-        const itemsById = Object.assign({}, state.itemsById);
-        const items = state.items.concat(action.data._items.map((item) => {
-            if (itemsById[item._id]) {
-                return;
-            }
-
-            itemsById[item._id] = item;
-            return item._id;
-        }).filter((_id) => _id));
-        return {...state, items, itemsById, isLoading: false};
-    }
-
-    case SET_STATE:
-        return Object.assign({}, action.state || initialState);
-
-    case RENDER_MODAL:
-    case CLOSE_MODAL:
-        return {...state, modal: modalReducer(state.modal, action)};
 
     case INIT_DATA: {
         const navigations = get(action, 'agendaData.navigations', []);
@@ -290,147 +225,8 @@ export default function agendaReducer(state = initialState, action) {
         };
     }
 
-    case ADD_TOPIC:
-        return {
-            ...state,
-            topics: state.topics.concat([action.topic]),
-        };
-
-    case TOGGLE_SELECTED:
-        return {
-            ...state,
-            selectedItems: toggleValue(state.selectedItems, action.item),
-        };
-
-    case SELECT_ALL:
-        return {
-            ...state,
-            selectedItems: state.items.concat(),
-        };
-
-    case SELECT_NONE:
-        return {
-            ...state,
-            selectedItems: [],
-        };
 
     case SELECT_DATE:
-        return {
-            ...state,
-            selectedItems: [],
-            activeDate: action.dateString,
-            activeGrouping: action.grouping || 'day',
-        };
-
-    case SHARE_ITEMS: {
-        const itemsById = updateItemActions(state, action.items, 'shares');
-
-        return {
-            ...state,
-            itemsById
-        };
-    }
-
-    case DOWNLOAD_ITEMS: {
-        const itemsById = updateItemActions(state, action.items, 'downloads');
-
-        return {
-            ...state,
-            itemsById
-        };
-    }
-
-    case COPY_ITEMS: {
-        const itemsById = updateItemActions(state, action.items, 'copies');
-
-        return {
-            ...state,
-            itemsById
-        };
-    }
-
-    case PRINT_ITEMS: {
-        const itemsById = updateItemActions(state, action.items, 'prints');
-
-        return {
-            ...state,
-            itemsById
-        };
-    }
-
-    case BOOKMARK_ITEMS: {
-        const itemsById = Object.assign({}, state.itemsById);
-        const bookmarkedItems = state.bookmarkedItems || [];
-
-        const missing = action.items.filter((item) => {
-            itemsById[item] = Object.assign({}, itemsById[item]);
-            itemsById[item].bookmarks = (itemsById[item].bookmarks || []).concat([state.user]);
-            return bookmarkedItems.indexOf(item) === -1;
-        });
-
-        return {
-            ...state,
-            itemsById,
-            bookmarkedItems: bookmarkedItems.concat(missing),
-        };
-    }
-
-    case REMOVE_BOOKMARK: {
-        const itemsById = Object.assign({}, state.itemsById);
-        const bookmarkedItems = state.bookmarkedItems || [];
-
-        const bookmarks = action.items.filter((item) => {
-            itemsById[item] = Object.assign({}, itemsById[item]);
-            itemsById[item].bookmarks = (itemsById[item].bookmarks || []).filter((val) => val !== state.user);
-            return bookmarkedItems.indexOf(item) === -1;
-        });
-
-        return {
-            ...state,
-            itemsById,
-            bookmarkedItems: bookmarks,
-        };
-    }
-
-    case SET_NEW_ITEMS_BY_TOPIC: {
-        const newItemsByTopic = Object.assign({}, state.newItemsByTopic);
-        action.data.topics.map((topic) => {
-            const previous = newItemsByTopic[topic] || [];
-            newItemsByTopic[topic] = previous.concat([action.data.item]);
-        });
-
-        let itemsById = state.itemsById;
-        if (get(action.data, 'item._id') && state.itemsById[action.data.item._id]) {
-            itemsById = Object.assign({}, itemsById);
-            itemsById[action.data.item._id] = action.data.item;
-        }
-
-        return {
-            ...state,
-            itemsById,
-            newItemsByTopic,
-        };
-    }
-
-    case REMOVE_NEW_ITEMS: {
-        const newItemsByTopic = Object.assign({}, state.newItemsByTopic);
-        newItemsByTopic[action.data] = null;
-        return {
-            ...state,
-            newItemsByTopic
-        };
-    }
-
-    case SET_NEW_ITEMS: {
-        const newItems = action.data._items.filter((item) => !item.nextversion && !state.itemsById[item._id]).map((item) => item._id);
-        const newItemsData = action.data;
-        return {
-            ...state,
-            newItems,
-            newItemsData,
-        };
-    }
-
     case TOGGLE_TOPIC:
     case TOGGLE_NAVIGATION:
     case TOGGLE_FILTER:
@@ -439,11 +235,6 @@ export default function agendaReducer(state = initialState, action) {
     case SET_VIEW:
         return {...state, agenda: _agendaReducer(state.agenda, action)};
 
-    case START_LOADING:
-        return {...state, isLoading: true};
-
-    case SET_TOPICS:
-        return {...state, topics: action.topics};
 
     default:
         return state;
